requirements.txt
openai 
openai-whisper 
python-dotenv 
pynput 
tabulate 
sounddevice 
soundfile 
numpy 
pandas 
langchain 
langchain-openai 
langchain-experimental
